+++ gangland-features/cops-n-crooks/src/main/java/me/luckyraven/copsncrooks/events/trader/TraderSellRequestEvent.java
@@ -9,6 +9,7 @@ import org.bukkit.event.HandlerList;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 
@@ -18,14 +19,15 @@ public class TraderSellRequestEvent extends TraderEvent implements Cancellable {
 	private static final HandlerList HANDLERS = new HandlerList();
 
 	private final List<ItemStack> offeredItems;
-	private final double          finalOffer;
+	private final BigDecimal      finalOffer;
 
 	@Setter
 	private boolean cancelled;
 	@Setter
 	private String  reason;
 
-	public TraderSellRequestEvent(Player player, TraderNpc trader, List<ItemStack> offeredItems, double finalOffer) {
+	public TraderSellRequestEvent(Player player, TraderNpc trader, List<ItemStack> offeredItems,
+	                              BigDecimal finalOffer) {
 		super(player, trader);
 		this.offeredItems = offeredItems == null ? Collections.emptyList() : List.copyOf(offeredItems);
 		this.finalOffer   = finalOffer;
