+++ gangland-ui/shop-api/src/main/java/me/luckyraven/shop/transaction/PaymentHandler.java
@@ -1,15 +1,17 @@
 package me.luckyraven.shop.transaction;
 
+import java.math.BigDecimal;
+
 /**
  * Minimal payment abstraction used by {@link ShopPurchaseService}. Integrations adapt this to whatever economy system
  * they run on top of (e.g. {@code EconomyHandler} in gangland-impl) — the shop-api itself stays economy-agnostic.
  */
 public interface PaymentHandler {
 
-	double getBalance();
+	BigDecimal getBalance();
 
-	void withdraw(double amount) throws PaymentException;
+	void withdraw(BigDecimal amount) throws PaymentException;
 
-	void deposit(double amount) throws PaymentException;
+	void deposit(BigDecimal amount) throws PaymentException;
 
 }
